requirements.txt
selenium
beautifulsoup4
pandas
psycopg2-binary